Fixed thirdparty integration tests
