+++ tests/Integration/TestCase.php
@@ -40,7 +40,6 @@ abstract class TestCase extends BaseTestCase {
 	public static function tear_down_after_class() {
 		parent::tear_down_after_class();
 
-		self::uninstallAll();
 
 		CapTrait::resetAdminCap();
 
@@ -55,6 +54,8 @@ abstract class TestCase extends BaseTestCase {
 				delete_transient( $transient );
 			}
 		}
+
+		self::uninstallAll();
 	}
 
 	public function set_up() {
