+++ tests/Integration/inc/Engine/CriticalPath/CriticalCSSSubscriber/deleteCpcss.php
@@ -2,6 +2,7 @@
 
 namespace WP_Rocket\Tests\Integration\inc\Engine\CriticalPath\CriticalCSSSubscriber;
 
+use WP_Rocket\Tests\Integration\DBTrait;
 use WP_Rocket\Tests\Integration\FilesystemTestCase;
 use WP_Rocket\Tests\Integration\ContentTrait;
 
@@ -13,7 +14,7 @@ use WP_Rocket\Tests\Integration\ContentTrait;
  * @group  vfs
  */
 class Test_DeleteCpcss extends FilesystemTestCase {
-	use ContentTrait;
+	use ContentTrait, DBTrait;
 
 	protected $path_to_test_data = '/inc/Engine/CriticalPath/CriticalCSSSubscriber/deleteCpcss.php';
 
@@ -23,6 +24,18 @@ class Test_DeleteCpcss extends FilesystemTestCase {
 
 	public static function wpSetUpBeforeClass( $factory ) {
 		self::$post_id = $factory->post->create();
+	}
+
+	public static function set_up_before_class()
+	{
+		self::installFresh();
+		parent::set_up_before_class();
+	}
+
+	public static function tear_down_after_class()
+	{
+		self::uninstallAll();
+		parent::tear_down_after_class();
 	}
 
 	public function set_up() {
